refactor lance, life drain, lightning staff, longsword, medicine, medic's kit, midas staff, mighty gauntlet, musket, and nature's caprice
