+++ source/gear/_gearDictionary.js
@@ -45,41 +45,16 @@ for (const file of [
 	"heat-weaken.js",
 	"herb-basket.js",
 	"illumination.js",
-	"lance-accelerating.js",
-	"lance-base.js",
-	"lance-bashing.js",
-	"lance-juggernauts.js",
-	"lance-taunting.js",
-	"lifedrain-base.js",
-	"lifedrain-fatiguing.js",
-	"lifedrain-furious.js",
-	"lifedrain-reapers.js",
-	"lifedrain-thirsting.js",
-	"lightningstaff-base.js",
-	"lightningstaff-disenchanting.js",
-	"lightningstaff-hexing.js",
-	"longsword-base.js",
-	"longsword-double.js",
-	"longsword-lethal.js",
-	"medicine-base.js",
-	"medicine-hastening.js",
-	"medicine-urgent.js",
-	"medicskit-base.js",
-	"medicskit-cleansing.js",
-	"medicskit-warning.js",
-	"midasstaff-accelerating.js",
-	"midasstaff-base.js",
-	"midasstaff-discounted.js",
-	"mightygauntlet-accurate.js",
-	"mightygauntlet-base.js",
-	"mightygauntlet-hearty.js",
-	"mightygauntlet-swift.js",
-	"musket-base.js",
-	"musket-discounted.js",
-	"musket-hunters.js",
-	"naturescaprice-accurate.js",
-	"naturescaprice-base.js",
-	"naturescaprice-hearty.js",
+	"lance.js",
+	"life-drain.js",
+	"lightning-staff.js",
+	"longsword.js",
+	"medicine.js",
+	"medics-kit.js",
+	"midas-staff.js",
+	"mighty-gauntlet.js",
+	"musket.js",
+	"natures-caprice.js",
 	"netlauncher-base.js",
 	"netlauncher-tormenting.js",
 	"netlauncher-thiefs.js",
